Added YAML Environment, Changed Settings, & Made Python Scripts
Added an updated version of the pandas.yml file to the repository.
This new environment contains a bunch of packages to help me in
Jupyter Lab.
Updated some PyCharm IDE settings.
Added an exercise.py file to test code for the first chapter.
The exercise.py file will use NumPy and SciPy, along with OOP,
to answer exercise.ipnyb.